arm: tegra: pluto: enable charger interrupt
passing of charger interrupt number from platform
data to max77665 charger driver.

(cherry picked from commit 9ecce79d6b2771f0275c25c35f5be49703219f54)
Reviewed-on: http://git-master/r/188544
Tested-by: Simone Willett <swillett@nvidia.com>

+++ drivers/power/max77665-charger.c
@@ -357,7 +357,7 @@ static irqreturn_t max77665_charger_irq_handler(int irq, void *data)
 {
 	struct max77665_charger *charger = data;
 	int ret;
-	uint8_t read_val;
+	uint32_t read_val;
 
 	ret = max77665_read_reg(charger, MAX77665_CHG_INT_OK, &read_val);
 	if (ret < 0) {
